src/main/java/com/mrunal/taskmanagement/repository/UserRepository.java
package com.mrunal.taskmanagement.repository;

import java.util.Optional;

import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.data.jpa.repository.Query;
import org.springframework.stereotype.Repository;

import com.mrunal.taskmanagement.entity.User;

@Repository
public interface UserRepository extends JpaRepository<User, Integer>{
	
	Optional<User> findByUsername(String username);
	
	@Query(value = "SELECT MAX(id) FROM user", nativeQuery = true)
	public int maxUserId();

}
